2019-18 snapshot
My simple a-star search worked fine for the first couple examples but
the fourth example is too complex, so I'm gonna have to do some
caching or problem space reduction or something. Just want to snapshot
this solution that works for the first couple examples in case I lose
my way.